renamed og-calendar and og-calendar-group
