--- src/components/og-calendar-group/og-calendar-group.scss
@@ -1,72 +1,0 @@
-:host {
-    --og-calendar__header-BorderWidth: 1px;
-    --og-calendar__header-BorderStyle: solid;
-    --og-calendar__header-BorderColor: var(--OG-COLOR-SHADE--100--20, #{$og-color-shade--100--20});
-    --og-calendar__header-Border: var(--og-calendar__header-BorderWidth) var(--og-calendar__header-BorderStyle) var(--og-calendar__header-BorderColor);
-
-    display: inline-block;
-    width: auto;
-}
-
-.calendar__container {
-    display: inline-block;
-
-    & + & {
-        margin-left: 32px;
-    }
-}
-
-.calendar__header {
-    display: flex;
-    text-align: center;
-    border-bottom: var(--og-calendar__header-Border);
-}
-
-.calender__header__main {
-    flex: 1;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.calender__header__prefix,
-.calender__header__suffix {
-    padding: 4px 8px;
-    min-width: 24px;
-}
-
-.calendar__main {
-    display: block;
-    padding-top: 8px;
-}
-
-.calendar__nav {
-    width: 24px;
-    height: 24px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-}
-
-.calendar__nav--hidden {
-    display: none;
-}
-
-.calendar__nav__icon {
-    width: 8px;
-}
-
-.calender__header__month {
-    display: inline-block;
-    font-size: 14px;
-    padding: 4px;
-    text-transform: uppercase;
-    font-weight: 500;
-}
-
-.calender__header__year {
-    display: inline-block;
-    // font-size: 12px;
-    padding: 4px;
-}
