+++ src/components/og-calendar/og-calendar.scss
@@ -1,83 +1,72 @@
 :host {
-    --og-calendar__day-Background: none;
-    --og-calendar__day-Background--hover: var(--OG-COLOR-PRIMARY--100--15, #{$og-color-primary--100--15});
-    --og-calendar__day-Background--active: var(--OG-COLOR-PRIMARY--100, #{$og-color-primary--100});
-    --og-calendar__day-Background--disabled: none;
+    --og-calendar__header-BorderWidth: 1px;
+    --og-calendar__header-BorderStyle: solid;
+    --og-calendar__header-BorderColor: var(--OG-COLOR-SHADE--100--20, #{$og-color-shade--100--20});
+    --og-calendar__header-Border: var(--og-calendar__header-BorderWidth) var(--og-calendar__header-BorderStyle) var(--og-calendar__header-BorderColor);
 
-    --og-calendar__day-Color: inherit;
-    --og-calendar__day-Color--hover: var(--OG-COLOR-PRIMARY--100--15, #{$og-color-primary--100--15});
-    --og-calendar__day-Color--active: var(--OG-COLOR-SHADE--0, #{$og-color-shade--0});
-    --og-calendar__day-Color--disabled: var(--OG-COLOR-SHADE--100--50, #{$og-color-shade--100--50});
-
-    --og-calendar__day__indicator-Background: var(--OG-COLOR-PRIMARY--40, #{$og-color-primary--40});
+    display: inline-block;
+    width: auto;
 }
 
-table {
-    border-spacing: 2px;
+.calendar__container {
+    display: inline-block;
+
+    & + & {
+        margin-left: 32px;
+    }
 }
 
-td {
-    line-height: 1;
-}
-
-th {
+.calendar__header {
+    display: flex;
     text-align: center;
-    text-transform: uppercase;
-    // font-size: 12px;
-
+    border-bottom: var(--og-calendar__header-Border);
 }
 
-.day {
-    background: var(--og-calendar__day-Background);
-    color: var(--og-calendar__day-Color);
-    text-align: center;
-    // font-size: 12px;
-    border-radius: 3px;
-    position: relative;
+.calender__header__main {
+    flex: 1;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.calender__header__prefix,
+.calender__header__suffix {
+    padding: 4px 8px;
+    min-width: 24px;
+}
+
+.calendar__main {
+    display: block;
+    padding-top: 8px;
+}
+
+.calendar__nav {
+    width: 24px;
+    height: 24px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
     cursor: pointer;
-    padding: 6px;
-
-
-    &::after {
-        position: absolute;
-        display: block;
-        content: '';
-        height: 2px;
-        top: auto;
-        right: 4px;
-        bottom: 2px;
-        left: 4px;
-    }
-
-    &:not(.day--disabled):not(.day--selected):hover {
-        --og-calendar__day-Background: var(--og-calendar__day-Background--hover, none);
-    }
 }
 
-.day--selected {
-    --og-calendar__day-Background: var(--og-calendar__day-Background--active, none);
-    --og-calendar__day-Color: var(--og-calendar__day-Color--active);
+.calendar__nav--hidden {
+    display: none;
 }
 
+.calendar__nav__icon {
+    width: 8px;
+}
 
+.calender__header__month {
+    display: inline-block;
+    font-size: 14px;
+    padding: 4px;
+    text-transform: uppercase;
+    font-weight: 500;
+}
 
-.week {
-    padding-right: 8px;
-    text-align: right;
+.calender__header__year {
+    display: inline-block;
     // font-size: 12px;
-    border-right: 1px solid var(--OG-COLOR-SHADE--100--20);
-}
-
-.day--disabled {
-    --og-calendar__day-Color: var(--og-calendar__day-Color--disabled);
-    --og-calendar__day-Background: var(--og-calendar__day-Background--disabled);
-    cursor: default;
-}
-
-.day--today {
-    font-weight: bold;
-}
-
-.day--highlight::after {
-    background: var(--og-calendar__day__indicator-Background);
+    padding: 4px;
 }
