+++ src/components/og-calendar/og-calendar.tsx
@@ -4,10 +4,10 @@
  * See LICENSE file at https://github.com/orgenic/orgenic-ui/blob/master/LICENSE
  **/
 
-import { h, Component, Prop, Event, EventEmitter } from '@stencil/core';
-import moment, { Moment } from 'moment';
-import { OgDateDecorator, OgCalendarDate } from './interfaces/og-calendar-date-decorator';
-import { CalendarUtils } from './utils/utils';
+import { h, Component, Prop, EventEmitter, Event } from '@stencil/core';
+import moment from 'moment';
+import { OgDateDecorator, OgCalendarSelectionType, OgCalendarDate } from '../og-internal-calendar/interfaces/og-calendar-date-decorator';
+import { CalendarUtils } from '../og-internal-calendar/utils/utils';
 import { loadMomentLocale, getDefaultLocale } from '../../utils/moment-locale-loader';
 
 @Component({
@@ -18,110 +18,136 @@ import { loadMomentLocale, getDefaultLocale } from '../../utils/moment-locale-lo
 export class OgCalendar {
     @Prop({ context: 'resourcesUrl' }) resourcesUrl!: string;
 
-    @Prop() year: number = new Date().getFullYear();
-    @Prop() month: number = new Date().getMonth();
+    @Prop({ reflectToAttr: true, mutable: true }) year: number = new Date().getFullYear();
+    @Prop({ reflectToAttr: true, mutable: true }) month: number = new Date().getMonth();
 
     @Prop() loc = getDefaultLocale();
 
     @Prop() showCalendarWeek: boolean = true;
+    @Prop() displayedMonths: number = 1;
     @Prop() dateDecorator: OgDateDecorator;
 
-    @Prop() selection: OgCalendarDate[];
+    @Prop() selectionType: OgCalendarSelectionType = 'single';
+    @Prop({ reflectToAttr: true, mutable: true }) selection: OgCalendarDate[] = [];
 
-    @Event() dateClicked: EventEmitter<Moment>;
+    @Event() dateClicked: EventEmitter<OgCalendarDate>;
+    @Event() selectionChanged: EventEmitter<OgCalendarDate[]>;
 
-    private internalMoment: Moment;
+    private internalMoment;
 
     async componentWillLoad() {
         await loadMomentLocale(this.loc, moment, this.resourcesUrl);
         this.internalMoment = moment();
     }
 
-    getFirstDayOfWeek() {
-        return this.internalMoment.startOf('week').day();
-    }
+    handleDateClick(event) {
+        event.cancelBubble = true;
 
-    getDayArray() {
-        return [0,1,2,3,4,5,6].map(d => {
-            return (d + this.getFirstDayOfWeek()) % 7;
-        });
-    }
-
-    getClasses(m: Moment) {
-        let result = 'day';
-        if (m.isSame(new Date(), "day")) {
-            result += ' day--today';
-        }
-
-        if (m.year() === this.year && m.month() === this.month) {
-            result += ' day--same-month';
-            if (this.selection && this.selection.find(s => CalendarUtils.compareCalendarDate2Moment(s, m))) {
-                result += ' day--selected';
-            }
-            if (this.dateDecorator) {
-                const dd = this.dateDecorator.getDateDecoration(m.clone());
-                if (dd.class) {
-                    result += ` ${dd.class}`;
+        const date = CalendarUtils.moment2CalendarDate(event.detail);
+        if (this.selectionType !== 'none') {
+            if (this.selectionType === 'single' || !this.selection) {
+                this.selection = [date];
+            } else {
+                // multi
+                // todo: implement selection types: range + multi-range
+                const index = this.selection.findIndex(s => CalendarUtils.compareCalendarDate(s, date));
+                if (index >= 0) {
+                    this.selection.splice(index, 1);
+                } else {
+                    this.selection.push(date);
                 }
+                this.selection = [...this.selection];
             }
-        } else {
-            result += ' day--disabled';
         }
-        return result;
+        this.dateClicked.emit(date);
+        this.selectionChanged.emit(this.selection);
     }
 
-    private setUpInternalMoment() {
-        this.internalMoment.locale(this.loc);
+    increaseMonth() {
+        this.internalMoment.add(1, 'M');
+        this.month = this.internalMoment.month();
+        this.year = this.internalMoment.year();
+    }
 
-        this.internalMoment.year(this.year);
-        this.internalMoment.month(this.month);
-        this.internalMoment.date(1);
-
-        const firstDayOfWeek = this.getFirstDayOfWeek();
-
-        if (this.internalMoment.day() < firstDayOfWeek) {
-            this.internalMoment.day(firstDayOfWeek - 7);
-        } else {
-            this.internalMoment.day(firstDayOfWeek);
-        }
+    decreaseMonth() {
+        this.internalMoment.subtract(1, 'M');
+        this.month = this.internalMoment.month();
+        this.year = this.internalMoment.year();
     }
 
     render() {
-        this.setUpInternalMoment();
+        this.internalMoment.locale(this.loc);
+        this.internalMoment.year(this.year);
+        this.internalMoment.month(this.month);
 
-        return [
-            <table>
-                <thead>
-                    <tr>
-                        { this.showCalendarWeek && <th></th> }
-                        {
-                            this.getDayArray().map(d => {
-                                return <th>{ this.internalMoment.day(d).format('dd') }</th>;
-                            })
-                        }
-                    </tr>
-                </thead>
-                <tbody>
-                    {
-                        [0,1,2,3,4,5].map(() => {
-                            return (<tr>
-                                { this.showCalendarWeek && <td class="week">{ this.internalMoment.week() }</td> }
-                                {
-                                    [0,1,2,3,4,5,6].map(() => {
-                                        const localM = this.internalMoment.clone();
-                                        this.internalMoment.add(1, 'd');
-                                        return <td
-                                                class={ this.getClasses(localM) }
-                                                onClick={ () => this.dateClicked.emit(localM) }>
-                                            { localM.date() }
-                                        </td>;
-                                    })
-                                }
-                            </tr>)
-                        })
-                    }
-                </tbody>
-            </table>
-        ];
+        const localM = this.internalMoment.clone();
+        const result = [];
+        for (let i = 0; i < this.displayedMonths; i++) {
+            result.push(
+                <div class="calendar__container">
+                    <div class="calendar__header">
+                        <div class="calender__header__prefix">
+                            <span class={ 'calendar__nav' + (i > 0 ? ' calendar__nav--hidden' : '') } onClick={ () => this.decreaseMonth() }>
+                                <svg
+                                    class={ 'calendar__nav__icon' }
+                                    version="1.1"
+                                    xmlns="http://www.w3.org/2000/svg"
+                                    xmlnsXlink="http://www.w3.org/1999/xlink"
+                                    viewBox="0 0 14 24"
+                                    preserveAspectRatio="none"
+                                >
+                                    <polyline
+                                        class="og-"
+                                        points="12,2 2,12 12,24"
+                                        fill="none"
+                                        stroke="currentColor"
+                                        stroke-width="2"
+                                        stroke-linecaps="round"
+                                        stroke-linejoin="round"
+                                    />
+                                </svg>
+                            </span>
+                        </div>
+                        <div class="calender__header__main">
+                            <span class="calender__header__month">{ localM.format('MMM') }</span><span class="calender__header__year">{ localM.year() }</span>
+                        </div>
+                        <div class="calender__header__suffix">
+                            <div class={ 'calendar__nav' + (i < this.displayedMonths - 1 ? ' calendar__nav--hidden' : '') } onClick={ () => this.increaseMonth() }>
+                                <svg
+                                    class={ 'calendar__nav__icon' }
+                                    version="1.1"
+                                    xmlns="http://www.w3.org/2000/svg"
+                                    xmlnsXlink="http://www.w3.org/1999/xlink"
+                                    viewBox="0 0 14 24"
+                                    preserveAspectRatio="none"
+                                >
+                                    <polyline
+                                        class="og-"
+                                        points="2,2 12,12 2,24"
+                                        fill="none"
+                                        stroke="currentColor"
+                                        stroke-width="2"
+                                        stroke-linecaps="round"
+                                        stroke-linejoin="round"
+                                    />
+                                </svg>
+                            </div>
+                        </div>
+                    </div>
+                    <og-internal-calendar
+                        class="calendar__main"
+                        month={ localM.month() }
+                        year={ localM.year() }
+                        loc={ this.loc }
+                        showCalendarWeek={ this.showCalendarWeek }
+                        dateDecorator={ this.dateDecorator }
+                        selection={ this.selection }
+                        onDateClicked={ (e) => this.handleDateClick(e) }>
+                    </og-internal-calendar>
+                </div>
+            );
+            localM.add(1, 'M');
+        }
+        return result;
     }
 }
