+++ src/index.html
@@ -166,9 +166,9 @@
             <og-card name="Calendar">
                 <div>
                     <h4>Calendar Group (2 months): de</h4>
-                    <og-calendar-group month="4" year="2019" loc="de" displayed-months="2" show-calendar-week="false" id="groupCalendar1"></og-calendar-group>
+                    <og-calendar month="4" year="2019" loc="de" displayed-months="2" show-calendar-week="false" id="groupCalendar1"></og-calendar>
                     <h4>Calendar Group (1 months): en</h4>
-                    <og-calendar-group month="4" year="2019" loc="en"></og-calendar-group>
+                    <og-calendar month="4" year="2019" loc="en"></og-calendar>
                 </div>
             </og-card>
             <script>
